+++ kedro_airflow_k8s/operators/node_pod.py
@@ -2,7 +2,7 @@
 Module contains Apache Airflow operator that creates k8s pod for execution of
 kedro node.
 """
-
+import logging
 from typing import Dict, Optional
 
 from airflow.kubernetes.pod_generator import PodGenerator
@@ -97,6 +97,15 @@ class NodePodOperator(KubernetesPodOperator):
             pod_template_file=self.minimal_pod_template,
             node_selector=node_selector_labels,
         )
+
+    def execute(self, context):
+        """
+        Executes task in pod with provided configuration (super implementation used).
+        :param context:
+        :return:
+        """
+        logging.info(self.create_pod_request_obj())
+        return super().execute(context)
 
     @staticmethod
     def create_resources(