+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+-   Logging added to operators
+
 ## [0.4.0] - 2021-04-20
 
 -   Support of S3 as DAG destination